requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.27
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.5
typing-extensions==4.13.2